refactor(frontend): refactores frontend file structure and convention (WEB-28)
